+++ rest/src/main/java/org/sing_group/mtc/rest/resource/route/ManagerWithResourcesRestPathBuilder.java
@@ -32,7 +32,7 @@ public class ManagerWithResourcesRestPathBuilder extends ManagerRestPathBuilder 
     return new InstitutionRestPathBuilder(this.builder);
   }
   
-  public InstitutionRestPathBuilder institution(int id) {
+  public InstitutionRestPathBuilder institution(long id) {
     return new InstitutionRestPathBuilder(this.builder, id);
   }
